Clean up some timeout arg warnings
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@87006 13f79535-47bb-0310-9956-ffa450edef68

+++ server/connection.c
@@ -208,7 +208,7 @@ void ap_lingering_close(conn_rec *c)
         if (timeout >= MAX_SECS_TO_LINGER) break;
 
         /* figure out the new timeout */
-        timeout = (MAX_SECS_TO_LINGER - timeout) * APR_USEC_PER_SEC;
+        timeout = (int)((MAX_SECS_TO_LINGER - timeout) * APR_USEC_PER_SEC);
     }
 
     apr_close_socket(c->client_socket);
